extract more fcs into notes lib
